binge-plus-gcp-manual-setup/requirements.txt
google-cloud-core==2.3.3
google-cloud-storage==2.9.0
google-cloud-iam==2.12.2
google-cloud-resource-manager==1.6.3
google-auth==2.22.0
protobuf>=4.21.0,<5.0.0
setuptools>=65.5.1
google-api-python-client